Add DigitalOcean base class with DigitalOceanCloud and AMDDevCloud sub classes
