+++ src/dstack/_internal/core/backends/amddevcloud/__init__.py
@@ -1,0 +1,1 @@
+# This package contains the implementation for the AMDDevCloud backend.
