+++ src/dstack/_internal/core/backends/amddevcloud/compute.py
@@ -1,0 +1,5 @@
+from dstack._internal.core.backends.digitalocean_base.compute import BaseDigitalOceanCompute
+
+
+class AMDDevCloudCompute(BaseDigitalOceanCompute):
+    pass
