+++ src/dstack/_internal/core/backends/digitalocean_base/__init__.py
@@ -1,0 +1,1 @@
+# This package contains the base classes for DigitalOcean and AMDDevCloud backends.
