+++ src/dstack/_internal/core/backends/digitalocean_base/api_client.py
@@ -3,30 +3,20 @@ from typing import Any, Dict, List, Optional
 import requests
 
 from dstack._internal.core.backends.base.configurator import raise_invalid_credentials_error
-from dstack._internal.core.errors import BackendRateLimitExceededError
+from dstack._internal.core.errors import NoCapacityError
 from dstack._internal.utils.logging import get_logger
 
 logger = get_logger(__name__)
 
-# DigitalOcean API endpoints
-STANDARD_CLOUD_API_URL = "https://api.digitalocean.com/v2"
-AMD_CLOUD_API_URL = "https://api-amd.digitalocean.com/v2"
-
 
 class DigitalOceanAPIClient:
-    def __init__(self, api_key: str, flavor: str = "standard"):
+    def __init__(self, api_key: str, api_url: str):
         self.api_key = api_key
-        self.flavor = flavor
-        self.base_url = self._get_base_url()
-
-    def _get_base_url(self) -> str:
-        if self.flavor == "amd":
-            return AMD_CLOUD_API_URL
-        return STANDARD_CLOUD_API_URL
+        self.base_url = api_url
 
     def validate_api_key(self) -> bool:
         try:
-            response = self._make_request("GET", "/account")
+            response = self._make_request("GET", "/v2/account")
             response.raise_for_status()
             return True
         except requests.HTTPError as e:
@@ -38,13 +28,13 @@ class DigitalOceanAPIClient:
             raise e
 
     def list_ssh_keys(self) -> List[Dict[str, Any]]:
-        response = self._make_request("GET", "/account/keys")
+        response = self._make_request("GET", "/v2/account/keys")
         response.raise_for_status()
         return response.json()["ssh_keys"]
 
     def create_ssh_key(self, name: str, public_key: str) -> Dict[str, Any]:
         payload = {"name": name, "public_key": public_key}
-        response = self._make_request("POST", "/account/keys", json=payload)
+        response = self._make_request("POST", "/v2/account/keys", json=payload)
         response.raise_for_status()
         return response.json()["ssh_key"]
 
@@ -58,17 +48,19 @@ class DigitalOceanAPIClient:
         return ssh_key["id"]
 
     def create_droplet(self, droplet_config: Dict[str, Any]) -> Dict[str, Any]:
-        response = self._make_request("POST", "/droplets", json=droplet_config)
+        response = self._make_request("POST", "/v2/droplets", json=droplet_config)
+        if response.status_code == 422:
+            raise NoCapacityError(response.json()["message"])
         response.raise_for_status()
         return response.json()["droplet"]
 
     def get_droplet(self, droplet_id: str) -> Dict[str, Any]:
-        response = self._make_request("GET", f"/droplets/{droplet_id}")
+        response = self._make_request("GET", f"/v2/droplets/{droplet_id}")
         response.raise_for_status()
         return response.json()["droplet"]
 
     def delete_droplet(self, droplet_id: str) -> None:
-        response = self._make_request("DELETE", f"/droplets/{droplet_id}")
+        response = self._make_request("DELETE", f"/v2/droplets/{droplet_id}")
         if response.status_code == 404:
             logger.debug("DigitalOcean droplet %s not found", droplet_id)
             return
@@ -90,8 +82,4 @@ class DigitalOceanAPIClient:
             json=json,
             timeout=timeout,
         )
-
-        if response.status_code == 429:
-            raise BackendRateLimitExceededError("API rate limit exceeded.")
-
         return response
