+++ src/dstack/_internal/core/backends/digitalocean_base/backend.py
@@ -1,0 +1,5 @@
+from dstack._internal.core.backends.base.backend import Backend
+
+
+class BaseDigitalOceanBackend(Backend):
+    pass
